Integrate external packages
I did some modification for the use of keyboard lib. Instead of using the key 'Enter' to pass if the user does not desire a hint, I chosed the key 'P'. I decided 'p' cuz the problem was that keyboard listenner was taking in consideration when you press 'Enter' to validate an input, so each time the user press 'Enter' to validate a guesses the keyboard listenner keep recording this as an input and will considerate that 'Enter' is pressed.
I try to release the key each moment is pressed but does not work as well.